requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.5
python-dotenv==1.0.0
pydantic==2.5.2
gunicorn==21.2.0
openai
instaloader
azure-storage-blob
mysql-connector-python
passlib[bcrypt]==1.7.4
